backend/requirements.txt
aiohttp==3.9.4
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
colorama==0.4.6
frozenlist==1.4.1
gtfs-realtime-bindings==1.0.0
idna==3.7
loguru==0.7.2
multidict==6.0.5
protobuf==5.26.1
python-dotenv==1.0.1
pytz==2024.1
win32-setctime==1.1.0
yarl==1.9.4
